+++ client/src/components/views/Chatroom/Chatroom.js
@@ -1,10 +1,10 @@
 import React, {useState, useEffect, useRef} from 'react';
 import {useSelector, useDispatch} from 'react-redux';
 import {Row, Col, Input} from 'antd';
-import './Section/RecordButton.css';
+import './Section/Shared/RecordButton.css';
 import './Chatroom.css'
-import Scenario from './Section/Scenario';
-import AudioList from './Section/AudioList';
+import Scenario from './Section/Client/Scenario';
+import AudioList from './Section/Shared/AudioList';
 import AudioRecordingScreen from './Section/Sub-container/AudioRecordingScreen'
 import {getRoom} from '../../../_actions/chatroom_actions'
 import TextChatScreen from './Section/Sub-container/TextChatScreen';
@@ -20,13 +20,15 @@ export default function Chatroom(props) {
   const message = useSelector(state => state.message);
   let userID = user.userData ? user.userData._id : "";
   let username = user.userData ? user.userData.name : "";
-  const [userRole, setUserRole] = useState("");
-  const [audioHistory, setAudioHistory] = useState([]);
+  const [ userRole, setUserRole ] = useState("");
+  const [ audioHistory, setAudioHistory ] = useState([]);
   const [ scenario, setScenario ] = useState([]);
   const [ progress, setProgress ] = useState([]);
 
   const dispatch = useDispatch();
 
+  // as they say, there's some problem with setState that I need to clean up so I'll just drop a bomb here as a mark
+  // vvvvv Flood gate to make sure dispatch is fired only once.
   if(userRole === "") {
     dispatch(getRoom(chatroomID))
     .then(async (response) => {
