+++ client/src/components/views/Chatroom/Section/Shared/AudioList.js
@@ -1,5 +1,7 @@
 import React, {useState} from 'react'
 import {Affix, Col, Row} from "antd";
+// import React from 'react'
+// import { Col, Row } from "antd";
 
 export default function AudioList(props) {
 
@@ -20,10 +22,12 @@ export default function AudioList(props) {
   return (
       <div style={{display: "flex", flexDirection: "column", justifyContent: "space-between", height: "100%"}}
            ref={setContainer}>
+       {/* <div style={{display: "flex", flexDirection: "column", justifyContent: "space-between", height: "100%"}}> */}
         <Row style={{fontWeight: 'bold', border: "1px solid white", flexGrow: '1',backgroundColor:"white"}}>
           <Col span={24} style={{textAlign: "center",fontsize:"18px"}}>Lịch sử hội thoại</Col>
+          {/* Empty affix will cause the system to fire a long-ass warning */}
           <Affix target={() => container}>
-
+            
           </Affix>
           {showAudio}
         </Row>
